--- packages/amplify-category-storage/src/headless-interface/v1/schemas/addStorage.schema.json
@@ -1,142 +1,0 @@
-{
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "definitions": {
-        "DynamoDBServiceConfiguration": {
-            "description": "Placeholder definition for DynamoDB config",
-            "properties": {
-                "placeholder": {
-                    "type": "string"
-                },
-                "serviceName": {
-                    "enum": [
-                        "dynamoDB"
-                    ],
-                    "type": "string"
-                }
-            },
-            "required": [
-                "placeholder",
-                "serviceName"
-            ],
-            "type": "object"
-        },
-        "LambdaTriggerConfig": {
-            "description": "Lambda function that runs on bucket change",
-            "properties": {
-                "mode": {
-                    "enum": [
-                        "existing",
-                        "new"
-                    ],
-                    "type": "string"
-                },
-                "name": {
-                    "type": "string"
-                }
-            },
-            "required": [
-                "mode",
-                "name"
-            ],
-            "type": "object"
-        },
-        "PermissionGroups": {
-            "additionalProperties": {
-                "items": {
-                    "enum": [
-                        "CREATE",
-                        "DELETE",
-                        "READ",
-                        "UPDATE"
-                    ],
-                    "type": "string"
-                },
-                "type": "array"
-            },
-            "description": "Permissions for Cognito user groups",
-            "type": "object"
-        },
-        "S3Permissions": {
-            "description": "Permissions that should be applied to the bucket",
-            "properties": {
-                "auth": {
-                    "description": "Permissions for authenticated users",
-                    "items": {
-                        "enum": [
-                            "CREATE",
-                            "DELETE",
-                            "READ",
-                            "UPDATE"
-                        ],
-                        "type": "string"
-                    },
-                    "type": "array"
-                },
-                "groups": {
-                    "$ref": "#/definitions/PermissionGroups"
-                },
-                "guest": {
-                    "description": "Permissions for unauthenticated users",
-                    "items": {
-                        "enum": [
-                            "CREATE",
-                            "DELETE",
-                            "READ",
-                            "UPDATE"
-                        ],
-                        "type": "string"
-                    },
-                    "type": "array"
-                }
-            },
-            "type": "object"
-        },
-        "S3ServiceConfiguration": {
-            "properties": {
-                "bucketName": {
-                    "description": "Globally unique bucket name",
-                    "type": "string"
-                },
-                "lambdaTrigger": {
-                    "$ref": "#/definitions/LambdaTriggerConfig"
-                },
-                "permissions": {
-                    "$ref": "#/definitions/S3Permissions"
-                },
-                "serviceName": {
-                    "enum": [
-                        "s3"
-                    ],
-                    "type": "string"
-                }
-            },
-            "required": [
-                "bucketName",
-                "permissions",
-                "serviceName"
-            ],
-            "type": "object"
-        }
-    },
-    "properties": {
-        "serviceConfiguration": {
-            "anyOf": [
-                {
-                    "$ref": "#/definitions/S3ServiceConfiguration"
-                },
-                {
-                    "$ref": "#/definitions/DynamoDBServiceConfiguration"
-                }
-            ]
-        },
-        "version": {
-            "type": "number"
-        }
-    },
-    "required": [
-        "serviceConfiguration",
-        "version"
-    ],
-    "type": "object"
-}
-
