+++ packages/amplify-category-storage/tsconfig.json
@@ -1,9 +1,13 @@
 {
   "extends": "../../tsconfig.base.json",
+  "include": ["**/*.js", "**/supported-services.json", "**/*.ts"],
   "compilerOptions": {
     "outDir": "./lib",
     "rootDir": "./src",
     "allowJs": true,
     "strict": false // converted from js
-  }
+  },
+  "references": [
+    {"path": "../amplify-util-headless-input"},
+  ]
 }
