+++ packages/amplify-util-headless-input/package.json
@@ -4,8 +4,6 @@
   "description": "Logic for validating objects against JSON-schema specs and performing version upgrades when necessary / possible",
   "main": "lib/index.js",
   "scripts": {
-    "test": "jest",
-    "test-ci": "jest --ci",
     "build": "tsc",
     "clean": "rimraf ./lib && rm tsconfig.tsbuildinfo"
   },
